Fix missing day in March
I do wonder if this has something to do with February having
an irregular number of days but the fix would suggest otherwise.

Instead of updating a `currentDay` variable I always calculate the
first day of the week based on the first visible day of the month
because it's less confusing and as long as I set the first date to
the start of the day, the bugs in #7 and #5 are not present which
is good enough for me.

refs #7